requirements.txt
numpy==1.19.5
keras
tensorflow-cpu
h5py
pillow
mtcnn
matplotlib
flask
opencv-contrib-python-headless
gunicorn
